xcms: use size_t for pointer offsets passed to strncmp
instead of converting to int and back

Fixes clang warnings of the form:
HVC.c:190:43: warning: implicit conversion changes signedness: 'int' to
      'unsigned long' [-Wsign-conversion]
          if (strncmp(spec, _XcmsTekHVC_prefix, n) != 0) {
              ~~~~~~~

+++ src/xcms/HVC.c
@@ -176,13 +176,13 @@ TekHVC_ParseString(
  *		XcmsSuccess if valid.
  */
 {
-    int n;
+    size_t n;
     char *pchar;
 
     if ((pchar = strchr(spec, ':')) == NULL) {
 	return(XcmsFailure);
     }
-    n = (int)(pchar - spec);
+    n = (size_t)(pchar - spec);
 
     /*
      * Check for proper prefix.
